backend/requirements.txt
fastapi
uvicorn[standard]
openai-agents
openai
yfinance
matplotlib
pandas
numpy
requests
python-dotenv
pydantic
beautifulsoup4
lxml
